mais counter is now a 16bits value
svn path=/trunk/iCub/firmware/; revision=30056

+++ motorControllerDsp56f807/common_source_code/can_interface.c
@@ -44,7 +44,7 @@ extern canmsg_t can_fifo[];
 extern Int16 write_p;
 extern Int16 read_p;
 
-extern UInt8 mais_counter;
+extern UInt16 mais_counter;
 canmsg_t _canmsg;					// buffer to prepare messages for send 
 
 byte  can_channels = 0;
